backend/requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.1.1
reportlab==4.4.10
sqlparse==0.5.5
